Added prepareCloneUploadFromUrl that also fetches the metadata. (#5)
This avoids problems from misspecification of the getProjectMetadata
parameters; only advanced developers need to deal with that. The README
instructions also simplify accordingly.

We also renamed prepareCloneUpload to prepareCloneUploadFromMetadata,
which accepts a project/version pair as positional arguments. This is
needed as the ID can't be inferred from _extra if we have raw metadata.